WIP: Start rewriting in TS Start rewriting the backend in TS
