+++ backend/models/BackendError.ts
@@ -1,0 +1,4 @@
+import * as t from 'io-ts';
+export const BackendError = t.partial({
+  error: t.string
+});
