+++ backend/models/Reason.ts
@@ -1,0 +1,3 @@
+import * as t from 'io-ts';
+
+export const Reason = t.string;
