+++ backend/src/redisWrapper.ts
@@ -1,10 +1,9 @@
 const redis = require('redis');
 const redisUrl = process.env.REDIS_URL || 'redis://localhost:6379';
-const cacheTtlSeconds = process.env.CACHE_TTL_SECONDS || 60;
 
 let redisClient;
 
-function instance() {
+const instance = () => {
   if (!redisClient) {
     // create a new redis client and connect to our local redis instance
     redisClient = redis.createClient({ url: redisUrl });
@@ -13,8 +12,8 @@ function instance() {
   }
 
   return redisClient;
-}
+};
 
 module.exports = {
-  instance: instance
+  instance
 };
