+++ package.json
@@ -6,7 +6,7 @@
     "start": "npm run backend",
     "dev": "ELM_APP_API_URL=http://localhost:4000 npm-run-all --parallel backend frontend",
     "frontend": "elm-app start",
-    "backend": "node ./backend/index.js",
+    "backend": "ts-node ./backend/index.ts",
     "build": "NODE_ENV=production elm-app build",
     "now-build": "NODE_ENV=production elm-app build",
     "deploy:now": "now --public ./build --local-config=../now.json && now alias --local-config=./now.json",
@@ -30,6 +30,7 @@
     "@koa/cors": "^2.2.3",
     "dotenv": "^6.2.0",
     "gtfs-realtime-bindings": "^0.0.4",
+    "io-ts": "1.8.2",
     "koa": "^2.6.2",
     "koa-bodyparser": "^4.2.1",
     "koa-cache-control": "^2.0.0",
@@ -38,6 +39,7 @@
     "redis": "^2.8.0",
     "request": "^2.88.0",
     "telegraf": "^3.26.0",
+    "ts-node": "8.0.2",
     "twit": "^2.2.11"
   },
   "devDependencies": {
@@ -48,7 +50,8 @@
     "husky": "^1.3.1",
     "lint-staged": "^8.1.0",
     "npm-run-all": "^4.1.5",
-    "prettier": "^1.16.0"
+    "prettier": "^1.16.0",
+    "typescript": "3.3.3"
   },
   "husky": {
     "hooks": {
