+++ tslint.json
@@ -1,0 +1,79 @@
+{
+  "rules": {
+    "adjacent-overload-signatures": true,
+    "curly": true,
+    "eofline": true,
+    "align": [true, "parameters"],
+    "class-name": true,
+    "indent": [true, "spaces"],
+    "no-angle-bracket-type-assertion": true,
+    "max-line-length": [true, 100],
+    "max-file-line-count": [true, 500],
+    "no-consecutive-blank-lines": [true, 1],
+    "member-access": false,
+    "no-trailing-whitespace": true,
+    "no-bitwise": true,
+    "no-debugger": true,
+    "prefer-const": true,
+    "no-empty-interface": true,
+    "no-string-throw": true,
+    "unified-signatures": true,
+    "space-before-function-paren": [
+      true,
+      {
+        "anonymous": "never",
+        "named": "never",
+        "asyncArrow": "always",
+        "method": "never",
+        "constructor": "never"
+      }
+    ],
+    "prefer-method-signature": false,
+    "arrow-return-shorthand": [true, "multiline"],
+    "no-duplicate-variable": true,
+    "no-inferrable-types": false,
+    "no-var-keyword": true,
+    "variable-name": [
+      true,
+      "ban-keywords",
+      "check-format",
+      "allow-leading-underscore",
+      "allow-pascal-case"
+    ],
+    "no-empty": true,
+    "no-shadowed-variable": true,
+    "no-unused-expression": true,
+    "triple-equals": true,
+    "jsdoc-format": true,
+    "one-line": [true, "check-else", "check-whitespace", "check-open-brace"],
+    "trailing-comma": [false, { "multiline": "never", "singleline": "never" }],
+    "quotemark": [false, "single", "avoid-escape"],
+    "semicolon": [false, "always", "ignore-interfaces"],
+    "typedef-whitespace": [
+      true,
+      {
+        "call-signature": "nospace",
+        "index-signature": "nospace",
+        "parameter": "nospace",
+        "property-declaration": "nospace",
+        "variable-declaration": "nospace"
+      },
+      {
+        "call-signature": "onespace",
+        "index-signature": "onespace",
+        "parameter": "onespace",
+        "property-declaration": "onespace",
+        "variable-declaration": "onespace"
+      }
+    ],
+    "whitespace": [
+      true,
+      "check-branch",
+      "check-decl",
+      "check-operator",
+      "check-separator",
+      "check-type"
+    ],
+    "no-console": ["warning", { "allow": ["warn", "error"] }]
+  }
+}
